separate some schedule logic
